pyproject.toml
[project]
name = "ankinote-ai"
dynamic = ["version"]
description = "AI-powered Anki vocabulary card generator using Google's suite of APIs"
requires-python = ">=3.13"
license = { text = "MIT" }
readme = "README.md"
authors = [
    { name = "Chris Chen", email = "chrischen@ignity.xyz" }
]
keywords = ["anki", "flashcards", "vocabulary", "ai", "google-ai", "language-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Education",
]
dependencies = [
    "aiohttp>=3.13.2",
    "click>=8.2.0",
    "google-cloud-aiplatform>=1.133.0",
    "google-cloud-texttospeech>=2.31.0",
    "google-genai[aiohttp]>=1.40.0",
    "google-images-search>=1.4.7",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "shortuuid>=1.0.13",
]

[project.scripts]
anki-deck = "ankinote:main"

[tool.hatch.version]
path = "src/ankinote/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/ankinote"]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[dependency-groups]
dev = [
    "basedpyright>=1.29",
    "ruff>=0.12.7",
    "pytest>=8.4.1",
    "pytest-cov>=6.1.1",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.pyright]
include = ["src/ankinote"]
exclude = [
    ".ruff_cache",
    "**/__pycache__",
]
typeCheckingMode = "strict"

[tool.pytest.ini_options]
minversion = "8.4.1"
testpaths = ["tests"]
asyncio_mode = "auto"
